286 Merges useMutateProcesses into useMutateProcess
