--- packages/orchestrator-ui-components/src/hooks/ProcessesHooks/index.ts
@@ -1,2 +1,0 @@
-export * from './useMutateProcess';
-export * from './useMutateProcesses';
