+++ packages/orchestrator-ui-components/src/hooks/index.ts
@@ -1,7 +1,6 @@
-export * from './ProcessesHooks';
-
 export * from './useQueryWithGraphql';
 export * from './useEngineStatusQuery';
+export * from './useMutateProcess';
 export * from './useOrchestratorConfig';
 export * from './useOrchestratorTheme';
 export * from './useProcessStatusCountsQuery';
